python/packages/jumpstarter-driver-sdwire/pyproject.toml
[project]
name = "jumpstarter-driver-sdwire"
dynamic = ["version", "urls"]
description = ""
authors = [
  { name = "Nick Cao", email = "ncao@redhat.com" },
]
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
dependencies = [
    "jumpstarter",
    "jumpstarter-driver-opendal",
    "pyusb>=1.2.1",
    "pyudev>=0.24.3",
]

[dependency-groups]
dev = [
    "pytest>=8.3.2",
    "pytest-cov>=5.0.0",
]

[tool.hatch.metadata.hooks.vcs.urls]
Homepage = "https://jumpstarter.dev"
source_archive = "https://github.com/jumpstarter-dev/repo/archive/{commit_hash}.zip"

[tool.hatch.version]
source = "vcs"
raw-options = { 'root' = '../../'}

[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"
